+++ src/components/DetailsMetrics/DetailsMetrics.js
@@ -19,7 +19,6 @@ such restriction.
 */
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
-import classNames from 'classnames'
 
 import InvocationMetricCard from './IncvocationMetricCard'
 import MetricChart from '../MetricChart/MetricChart'
@@ -73,13 +72,6 @@ const DetailsMetrics = ({ selectedItem }) => {
   const generatedMetrics = useMemo(() => {
     return groupMetricByApplication(metrics, true)
   }, [metrics])
-
-  const invocationCardClassnames = classNames(
-    'metrics__card-invocations',
-    !isInvocationCardExpanded
-      ? 'metrics__card-invocations-header-hidden'
-      : 'metrics__card-invocations-header-visible'
-  )
 
   const calculateHistogram = useCallback((points, metric) => {
     const numberOfBins = 5
@@ -402,16 +394,14 @@ const DetailsMetrics = ({ selectedItem }) => {
                     )
                   } else {
                     return (
-                      <div className={invocationCardClassnames}>
-                        <InvocationMetricCard
-                          ref={invocationBodyCardRef}
-                          isInvocationCardExpanded={isInvocationCardExpanded}
-                          key={metric.id}
-                          metric={metric}
-                          previousTotalInvocation={previousTotalInvocation}
-                          selectedDate={selectedDate}
-                        />
-                      </div>
+                      <InvocationMetricCard
+                        ref={invocationBodyCardRef}
+                        isInvocationCardExpanded={isInvocationCardExpanded}
+                        key={metric.id}
+                        metric={metric}
+                        previousTotalInvocation={previousTotalInvocation}
+                        selectedDate={selectedDate}
+                      />
                     )
                   }
                 } else if (!metric.data) {
